Update Maya exporter
Modify the Maya exporter so that export will search all curve nodes
under the selected nodes recursively. This also fixes a python exception
bug when a non-curve type is selected. Now the code checks the node type
before attempting to convert it to curve node. Example models were
re-exported with the new exporter.