Change simulation UAV code to match previous style
I probably actually prefer the @mrgreywater version, but I'm changing it
back to the old style, which is also the style of the other pair of
CSSetUnorderedAccessViews calls. This is mostly just to make the diffs a
bit cleaner when diffing the separate @mrgreywater pull requests as a
single commit.