Release TressFX 3.1.0
Our biggest update in this release is a new order-independent
transparency (OIT) option we call "ShortCut". We've also addressed some
of the issues brought up by the community.

ShortCut is our new Order Independent Transparency (OIT) option. It's
inspired by the method presented by Eidos-Montréal and Hybrid
Transparency. Whereas our original method focused on the front k = 8 or
so layers of hair, ShortCut is good for cases when you can get away with
k = 2 or 3, and you're more concerned about memory usage.

This update also includes an alternative method for keeping the fur in
sync with the underlying animated mesh. This method has better
performance with ShortCut.